[HyeonJun] Foreground Service 오류 수정 및 구현 완료
- HomeFragment.kt : 기존에 Beacon Scanner를 직접 실행했었는데, 이 부분을 삭제했습니다.
- ForegroundService.kt : Beacon Scanner 기능을 이식했습니다. 이제 사용자가 앱을 종료해도 Scanning 동작이 종료되지 않습니다.
