+++ Application/Project_Footprint/app/src/main/AndroidManifest.xml
@@ -2,6 +2,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.haerokim.project_footprint">
 
+    <uses-feature
+        android:name="android.hardware.bluetooth_le"
+        android:required="true" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-feature android:name="android.hardware.location.network" />
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
@@ -24,7 +27,8 @@
         android:usesCleartextTraffic="true">
 
         <activity
-            android:name=".HomeActivity">
+            android:name=".HomeActivity"
+            android:launchMode="singleInstance">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
