+++ Application/Project_Footprint/app/src/main/res/layout/fragment_home.xml
@@ -29,7 +29,7 @@
     <androidx.appcompat.widget.AppCompatToggleButton
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/toggle_test"
+        android:id="@+id/scanning_mode_switch"
         android:text="ON"/>
 
 </LinearLayout>
