+++ Application/Project_Footprint/app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#6200EE</color>
-    <color name="colorPrimaryDark">#3700B3</color>
-    <color name="colorAccent">#03DAC5</color>
+    <color name="colorPrimary">#293263</color>
+    <color name="colorPrimaryDark">#000b38</color>
+    <color name="colorAccent">#565b91</color>
 </resources>
